backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "TwoCards Platform backend"
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
sqlmodel = "^0.0.22"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.28.1"
cryptography = "^44.0.1"
python-multipart = "^0.0.20"
uvicorn = "^0.34.0"
asyncpg = "^0.30.0"
aiosqlite = "^0.20.0"
pydantic-settings = "^2.7.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
httpx = {extras = ["cli"], version = "^0.28.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
